python/requirement.txt
fastapi
uvicorn
langchain
youtube-transcript-api
yt-dlp
fastapi
pydantic
langchain
langchain-google-genai
langchain-experimental
langchain-groq
llama-index
llama-index-llms-openrouter
llama-index-agent-openai
llama-index-core
llama-index-experimental
requests
beautifulsoup4
pandas
yt-dlp
faiss-cpu
langchain-community
langchain-text-splitter
langchain-vectorstores
langchain-retrievers
langchain-prompts
langchain-core
smtplib
ssl
email
langchain_google_genai
